Register service worker.
