+++ src/index.js
@@ -6,7 +6,7 @@ import AppRoute from './Routes/AppRoute';
 import { Provider } from 'react-redux';
 import store from './store/Configure';
 import { REQUEST_CONFIG_DATA } from './App/Actions';
-//import register from './registerServiceWorker';
+import * as serviceWorker from './serviceWorker';
 
 const {App} = ApplicationRoutes;
 const {Routes} = ApplicationRoutes;
@@ -41,4 +41,4 @@ const Morpheus = () => {
 
 ReactDOM.render(<Morpheus />, document.getElementById('root'));
 
-//register();
+serviceWorker.register();
